# openapi.yaml
openapi: 3.0.3
info:
  title: Resort Backend API (compat + core)
  version: 1.0.0
  description: "Machine-readable contract for the Resort Backend used by the frontend. Includes compat `/api` endpoints, admin/internal headers, and upload specs."
servers:
  - url: "{API_BASE}"
    description: "Runtime server (use VITE_API_URL or REACT_APP_API_URL as API_BASE)"
    variables:
      API_BASE:
        default: "http://localhost:8000"
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        capacity:
          type: integer
        price_per_night:
          type: number
        available:
          type: boolean
    Cottage:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        price:
          type: number
        capacity:
          type: integer
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
    Booking:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        status:
          type: string
        allocated_cottages:
          type: array
          items:
            $ref: '#/components/schemas/Cottage'
...
paths:
...
  /bookings/:
    get:
      summary: List bookings
      responses: { '200': { description: OK } }
    post:
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /bookings/{booking_id}:
    get:
      parameters: [{ name: booking_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      parameters: [{ name: booking_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: booking_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
  /bookings/guest/{guest_email}:
    get:
      parameters: [{ name: guest_email, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
...
  /api/cottages:
    get:
      summary: Compat cottages endpoint
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses: { '200': { description: OK } }
  /api/cottages/{cottage_id}:
    get:
      parameters: [{ name: cottage_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /api/bookings:
    post:
      summary: Create a booking
      tags:
        - bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - guest_name
                - guest_email
                - guests
                - check_in
                - check_out
              properties:
                guest_name:
                  type: string
                guest_email:
                  type: string
                  format: email
                guest_phone:
                  type: string
                guests:
                  type: integer
                  minimum: 1
                selected_cottages:
                  type: array
                  items:
                    type: string
                    description: MongoDB _id of preferred cottage
                check_in:
                  type: string
                  format: date
                check_out:
                  type: string
                  format: date
                allow_extra_beds:
                  type: boolean
                preferred_room_types:
                  type: array
                  items:
                    type: string
                price_breakdown:
                  type: object
                  additionalProperties: true
                special_requests:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  reference:
                    type: string
                  status:
                    type: string
                  allocated_cottages:
                    type: array
                    items:
                      type: string
                  price_breakdown:
                    type: object
                    additionalProperties: true
        '400':
          description: Validation error or no availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '409':
          description: Conflict (allocation race)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '503':
          description: DB unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
...
security: []